refactor: Update OpenRazer configuration for user 'esinger'
This commit updates the OpenRazer configuration in the NixOS repository for the user 'esinger'. It adds 'esinger' to the list of users for OpenRazer, enabling the hardware features for that user. The previous system package 'prismlauncher' is commented out, as it is no longer needed. This change aims to improve the customization and management of OpenRazer settings.