refactor: Update disk device configuration in modules/disko.nix
This commit updates the disk device configuration in the modules/disko.nix file. The `device` attribute is modified to allow for customization of the disk device path. This change improves the flexibility and usability of the disko module.

Note: This commit message is based on the provided code changes and recent repository commits.