refactor: Add NFS mount for /mnt/media in default.nix
This commit adds an NFS mount for the /mnt/media directory in the default.nix file. The device is set to "10.0.0.3:/mnt/stuff/media" and the fsType is set to "nfs". This change enhances the file system capabilities of the event-horizon host.

Note: This commit message is based on the provided code changes and recent repository commits.