refactor: Update virtualization services to use Docker instead of Podman
This commit updates the virtualization services in the default.nix and virtualization modules to use Docker instead of Podman. The `docker.enable` attribute is set to true, enabling Docker support. This change improves the compatibility and functionality of the virtualization services.

Note: This commit message is based on the provided code changes and recent repository commits.