refactor: Enable rootless Docker with socket variable in docker.nix
This commit enables rootless Docker with the socket variable in the docker.nix file. The `rootless.enable` attribute is set to true and `rootless.setSocketVariable` is also set to true. This change enhances the security and functionality of the Docker virtualization service.

Note: This commit message is based on the provided code changes and recent repository commits.