refactor: Update autoPrune attribute in docker.nix
This commit updates the `autoPrune` attribute in the `docker.nix` file. The attribute is modified to use a nested object with the `enable` attribute set to true. This change enhances the functionality of the Docker virtualization service.

Note: This commit message is based on the provided code changes and recent repository commits.